Ignoring errors is generally bad practice, so let's stop doing that.
Right now we are assuming the "main" functions (`CheckOfflineNUCs` and `CheckIdleMachines`) return at most one error, but in the future, it might make sense to allow more than one error. This might provide a better mechanism for handling Metricly API errors, as opposed to just logging them and moving on.