Renamed metricly/elements file to metricly/samples.
In spite of the fact that the first element of the URL these methods call is "element", I think it is a confusing label to give them. Additionally, I think "FetchMetricValue" is a more confusing function name to export than "FetchMetricSample" would be. I'm not making that change here, because I want this to be a minor version bump, and that would require a major bump.